src/ui/requirements.txt
mmh3
numpy
pandas
streamlit
st-annotated-text
Markdown
httpx
streamlit-chat
pydantic
# also farm-haystack without dependencies (only used for schema evaluation)